+++ recordlinkage/preprocessing/__init__.py
@@ -1,2 +1,5 @@
 from recordlinkage.preprocessing.cleaning import *
-from recordlinkage.preprocessing.encoding import *
+from recordlinkage.preprocessing.encoding import phonetic
+
+from recordlinkage.preprocessing.encoding import _list_phonetic_algorithms
+phonetic_algorithms = _list_phonetic_algorithms()